Move MapDBBlockingDatabase.Block to AbstractBlockingDatabase.Block to preserve DatabaseStatistics block size feature
(See larsga/Duke#150 for more information)
