+++ duke-mapdb/src/main/java/no/priv/garshol/duke/databases/MapDBBlockingDatabase.java
@@ -1,27 +1,21 @@
 
 package no.priv.garshol.duke.databases;
 
-import java.util.Map;
-import java.util.NavigableMap;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.io.File;
 import java.io.DataInput;
 import java.io.DataOutput;
+import java.io.File;
 import java.io.IOException;
 import java.io.Serializable;
+import java.util.Collection;
+import java.util.Map;
+import java.util.NavigableMap;
 
+import no.priv.garshol.duke.CompactRecord;
+import no.priv.garshol.duke.Property;
+import no.priv.garshol.duke.Record;
 import org.mapdb.DB;
 import org.mapdb.DBMaker;
 import org.mapdb.Serializer;
-
-import no.priv.garshol.duke.Record;
-import no.priv.garshol.duke.Property;
-import no.priv.garshol.duke.Database;
-import no.priv.garshol.duke.Configuration;
-import no.priv.garshol.duke.CompactRecord;
 
 // FIXME:
 //  - is a mapdb-based link database a good idea?
@@ -231,53 +225,6 @@ public class MapDBBlockingDatabase extends AbstractBlockingDatabase {
         .make();
     else
       return db.getTreeMap(name);
-  }
-
-  // --- BLOCK CONTAINER
-
-  public static class Block implements Serializable {
-    private int free;
-    private String[] ids;
-
-    public Block() {
-      this.ids = new String[10];
-    }
-
-    public Block(int free, String[] ids) {
-      this.free = free;
-      this.ids = ids;
-    }
-
-    public String[] getIds() {
-      return ids;
-    }
-
-    public void add(String id) {
-      if (free >= ids.length) {
-        String[] newids = new String[ids.length * 2];
-        for (int ix = 0; ix < ids.length; ix++)
-          newids[ix] = ids[ix];
-        ids = newids;
-      }
-      ids[free++] = id;
-    }
-
-    public void remove(String id) {
-      for (int ix = 0; ix < free; ix++) {
-        if (ids[ix].equals(id)) {
-          free--;
-          ids[ix] = ids[free];
-          // we don't need to null out the free cell in the array.
-          // reducing 'free' is sufficient.
-          return;
-        }
-      }
-      // FIXME: if we get here something's wrong. add a check?
-    }
-
-    public int size() {
-      return free;
-    }
   }
 
   // ----- SERIALIZERS
